PP-13526 Cleanup old team members code
