+++ src/controllers/user/phone-number/post.controller.js
@@ -1,10 +1,10 @@
 'use strict'
 
-const { response } = require('../../../utils/response')
-const userService = require('../../../services/user.service')
-const paths = require('../../../paths')
-const { invalidTelephoneNumber } = require('../../../utils/telephone-number-utils')
-const { validationErrors } = require('../../../utils/validation/field-validation-checks')
+const { response } = require('@utils/response')
+const userService = require('@services/user.service')
+const paths = require('@root/paths')
+const { invalidTelephoneNumber } = require('@utils/telephone-number-utils')
+const { validationErrors } = require('@utils/validation/field-validation-checks')
 
 module.exports = async function updatePhoneNumber (req, res, next) {
   const telephoneNumber = req.body.phone
@@ -15,7 +15,7 @@ module.exports = async function updatePhoneNumber (req, res, next) {
         phone: validationErrors.invalidTelephoneNumber
       }
     }
-    return response(req, res, 'team-members/edit-phone-number', pageData)
+    return response(req, res, 'user/edit-phone-number', pageData)
   }
 
   try {
