+++ src/utils/display-converter.js
@@ -28,14 +28,8 @@ const hideServiceNavTemplates = [
   'services/edit-service-name',
   'services/add-service',
   'services/select-org-type',
-  'merchant-details/merchant-details',
-  'merchant-details/edit-merchant-details',
-  'team-members/team-members',
-  'team-members/team-member-invite',
-  'team-members/team-member-details',
-  'team-members/team-member-profile',
-  'team-members/team-member-permissions',
-  'team-members/edit-phone-number',
+  'user/my-profile',
+  'user/edit-phone-number',
   'request-to-go-live/agreement',
   'request-to-go-live/choose-how-to-process-payments',
   'request-to-go-live/index',
