+++ src/utils/types/express/Message.ts
@@ -1,0 +1,9 @@
+
+type MessageState = 'success' | 'error' | 'info'
+
+export default interface Message {
+  state: MessageState
+  icon?: string
+  heading: string
+  body?: string
+}
