+++ src/utils/types/express/ServiceRequest.ts
@@ -1,25 +1,8 @@
-import type User from '@models/User.class'
 import type Service from '@models/Service.class'
 import type GatewayAccount from '@models/GatewayAccount.class'
-import type { Request } from 'express'
-import type StripeAccountSetup from '@models/StripeAccountSetup.class'
-import ClientSessionsCookie from '@utils/types/client-sessions/ClientSessionsCookie'
+import UserRequest from "@utils/types/express/UserRequest";
 
-type MessageState = 'success' | 'error' | 'info'
-
-interface Message {
-  state: MessageState
-  icon?: string
-  heading: string
-  body?: string
-}
-
-export default interface ServiceRequest<T = never> extends Request {
-  user: User
+export default interface ServiceRequest<T = never> extends UserRequest {
   service: Service
   account: GatewayAccount
-  gatewayAccountStripeProgress?: StripeAccountSetup
-  flash(type: string, message: Message): void
-  body: T,
-  session: ClientSessionsCookie
 }
