--- src/views/team-members/team-member-details.njk
@@ -1,70 +1,0 @@
-{% extends "../layout.njk" %}
-
-{% block pageTitle %}
-  Team members details — GOV.UK Pay
-{% endblock %}
-
-{% block beforeContent %}
-  {{ super() }}
-  {{
-    govukBackLink({
-    text: "See all team members",
-    href: teamMemberIndexLink,
-    classes: "link-back"
-    })
-  }}
-{% endblock %}
-
-{% block mainContent %}
-<div class="govuk-grid-column-two-thirds admin-list">
-  <span class="govuk-caption-l">Team members</span>
-  <h1 id="details-for" class="govuk-heading-l ">
-    Details for {{email}}
-  </h1>
-
-  <table class="govuk-table">
-    <caption class="govuk-table__caption govuk-visually-hidden">Member details</caption>
-    <tbody class="govuk-table__body">
-      <tr class="govuk-table__row">
-        <th class="govuk-table__header" scope="row">Email:</th>
-        <td class="govuk-table__cell" colspan="2" id="email">{{email}}</td>
-      </tr>
-      <tr class="govuk-table__row">
-        <th class="govuk-table__header" scope="row">Permission level:</th>
-        <td class="govuk-table__cell" id="role">{{role}}</td>
-        <td class="govuk-table__cell pay-text-align-right" id="edit-permissions-link">
-          <a class="govuk-link govuk-link--no-visited-state" id="edit-permissions" href="{{editPermissionsLink}}">Edit permissions</a>
-        </td>
-      </tr>
-    </tbody>
-  </table>
-
-  {% if permissions.users_service_delete %}
-  <p class="govuk-body">
-    <a class="govuk-link govuk-link--no-visited-state target-to-show--toggle" id="remove-team-member" href="#confirm-remove-user-dialog">Remove user</a>
-  </p>
-  <div id="confirm-remove-user-dialog" class="target-to-show">
-    <div class="govuk-error-summary govuk-!-margin-top-6" role="group" aria-labelledby="error-summary" tabindex="-1">
-      <h2 class="govuk-error-summary__title" id="error-summary">
-        Are you sure you want to remove {{email}}?
-      </h2>
-      <form id="remove-team-member-form" method="post" action="{{removeTeamMemberLink}}" novalidate>
-        <input id="csrf" name="csrfToken" type="hidden" value="{{csrf}}"/>
-
-        <div class="govuk-button-group">
-          {{govukButton({
-              text: "Yes, remove",
-              classes: "govuk-button--warning",
-              attributes: {
-                id: "remove-team-member-confirm"
-              }
-          })}}
-
-          <a href="#main" class="govuk-link target-to-show--cancel" id="remove-team-member-cancel">Cancel</a>
-        </div>
-      </form>
-    </div>
-  </div>
-  {% endif %}
-</div>
-{% endblock %}
