--- src/views/team-members/team-member-permissions.njk
@@ -1,203 +1,0 @@
-{% extends "../layout.njk" %}
-
-{% block pageTitle %}
-Edit team member permissions - GOV.UK Pay
-{% endblock %}
-
-{% block beforeContent %}
-  {{ super() }}
-  {{
-    govukBackLink({
-      text: "Back to team member profile",
-      href: teamMemberProfileLink,
-      classes: "link-back"
-    })
-  }}
-{% endblock %}
-
-{% block mainContent %}
-<div class="govuk-grid-column-full">
-  <span class="govuk-caption-l govuk-!-margin-top-6">Team members</span>
-  <h1 class="govuk-heading-l">
-    Edit permissions
-  </h1>
-</div>
-
-<div class="govuk-grid-column-two-thirds">
-  <dl class="read-only-fields">
-    <dt class="govuk-heading-s govuk-!-margin-bottom-0">Email address</dt>
-    <dd class="govuk-body govuk-!-margin-left-0" id="email">{{email}}</dd>
-  </dl>
-  <form id="role-update-form" method="post" action="{{editPermissionsLink}}" novalidate>
-    <input id="csrf" name="csrfToken" type="hidden" value="{{csrf}}"/>
-    {% if serviceHasAgentInitiatedMotoEnabled %}
-      {{
-        govukRadios({
-          idPrefix: 'role-input',
-          name: 'role-input',
-          fieldset: {
-            legend: {
-              text: 'Permission level',
-              classes: 'govuk-fieldset__legend--s'
-            }
-          },
-          items: [
-            {
-                value: admin.id,
-                id: "role-admin-input",
-                text: 'Administrator',
-                classes: 'govuk-label--s',
-                checked: admin.checked,
-                hint: {
-                  html: '<ul class="govuk-list pay-text-grey">
-                          <li>View transactions</li>
-                          <li>Refund payments</li>
-                          <li>Take telephone payments</li>
-                          <li>Manage settings</li>
-                        </ul>'
-                }
-            },
-            {
-                value: viewAndRefund.id,
-                id: "role-view-and-refund-input",
-                text: 'View and refund',
-                classes: 'govuk-label--s',
-                checked: viewAndRefund.checked,
-                hint: {
-                  html: '<ul class="govuk-list pay-text-grey">
-                          <li>View transactions</li>
-                          <li>Refund payments</li>
-                          <li>Cannot take telephone payments</li>
-                          <li>Cannot manage settings</li>
-                        </ul>'
-                }
-            },
-            {
-                value: view.id,
-                id: "role-view-input",
-                text: 'View only',
-                classes: 'govuk-label--s',
-                checked: view.checked,
-                hint: {
-                  html: '<ul class="govuk-list pay-text-grey">
-                          <li>View transactions</li>
-                          <li>Cannot refund payments</li>
-                          <li>Cannot take telephone payments</li>
-                          <li>Cannot manage settings</li>
-                        </ul>'
-                }
-            },
-            {
-                value: viewAndInitiateMoto.id,
-                id: "role-view-and-initiate-moto-input",
-                text: 'View and take telephone payments',
-                classes: 'govuk-label--s',
-                checked: viewAndInitiateMoto.checked,
-                hint: {
-                  html: '<ul class="govuk-list pay-text-grey">
-                          <li>View transactions</li>
-                          <li>Cannot refund payments</li>
-                          <li>Take telephone payments</li>
-                          <li>Cannot manage settings</li>
-                        </ul>'
-                }
-            },
-            {
-                value: viewRefundAndInitiateMoto.id,
-                id: "role-view-refund-and-initiate-moto-input",
-                text: 'View, refund and take telephone payments',
-                classes: 'govuk-label--s',
-                checked: viewRefundAndInitiateMoto.checked,
-                hint: {
-                  html: '<ul class="govuk-list pay-text-grey">
-                          <li>View transactions</li>
-                          <li>Refund payments</li>
-                          <li>Take telephone payments</li>
-                          <li>Cannot manage settings</li>
-                        </ul>'
-                }
-            }
-          ]
-        })
-      }}
-    {% else %}
-      {{
-        govukRadios({
-          idPrefix: 'role-input',
-          name: 'role-input',
-          fieldset: {
-            legend: {
-              text: 'Permission level',
-              classes: 'govuk-fieldset__legend--s'
-            }
-          },
-          items: [
-            {
-                value: admin.id,
-                id: "role-admin-input",
-                text: 'Administrator',
-                classes: 'govuk-label--s',
-                checked: admin.checked,
-                hint: {
-                  html: '<ul class="govuk-list pay-text-grey">
-                          <li>View transactions</li>
-                          <li>Refund payments</li>
-                          <li>Manage settings</li>
-                        </ul>'
-                }
-            },
-            {
-                value: viewAndRefund.id,
-                id: "role-view-and-refund-input",
-                text: 'View and refund',
-                classes: 'govuk-label--s',
-                checked: viewAndRefund.checked,
-                hint: {
-                  html: '<ul class="govuk-list pay-text-grey">
-                          <li>View transactions</li>
-                          <li>Refund payments</li>
-                          <li>Cannot manage settings</li>
-                        </ul>'
-                }
-            },
-            {
-                value: view.id,
-                id: "role-view-input",
-                text: 'View only',
-                classes: 'govuk-label--s',
-                checked: view.checked,
-                hint: {
-                  html: '<ul class="govuk-list pay-text-grey">
-                          <li>View transactions</li>
-                          <li>Cannot refund payments</li>
-                          <li>Cannot manage settings</li>
-                        </ul>'
-                }
-            }
-          ]
-        })
-      }}
-    {% endif %}
-
-    {{
-      govukWarningText({
-        text: 'This change affects both live and test environments',
-        iconFallbackText: 'Warning'
-      })
-    }}
-
-    {{
-      govukButton({
-        html: 'Save changes',
-        classes: 'button'
-      })
-    }}
-
-    <p class="govuk-body">
-      <a class="govuk-link govuk-link--no-visited-state" id="service-name-cancel-link" href="{{teamMemberIndexLink}}">
-        Cancel
-      </a>
-    </p>
-  </form>
-</div>
-{% endblock %}
